+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Detect network status changes in local-first services.
+
 ## [v0.1.0](https://github.com/NoelDeMartin/aerogel/releases/tag/v0.1.0) - 2025-05-30
 
 First release! No documentation yet, but _it's coming_.